# Conflicts:
#	aclk/legacy/Makefile.am
#	tests/Makefile.am
#	web/gui/dashboard/Makefile.am

+++ packaging/installer/netdata-uninstaller.sh
@@ -1,4 +1,4 @@
-#!/usr/bin/env bash
+#!/bin/sh
 #
 # This is the netdata uninstaller script
 #
@@ -10,6 +10,8 @@
 #
 # Author: Paweł Krupa <paulfantom@gmail.com>
 # Author: Pavlos Emm. Katsoulakis <paul@netdata.cloud>
+#
+# Next unused error code: R0005
 
 usage="$(basename "$0") [-h] [-f ] -- program to calculate the answer to life, the universe and everything
 
@@ -21,6 +23,7 @@ where:
 
 FILE_REMOVAL_STATUS=0
 ENVIRONMENT_FILE="/etc/netdata/.environment"
+# shellcheck disable=SC2034
 INTERACTIVITY="-i"
 YES=0
 while :; do
@@ -35,6 +38,7 @@ while :; do
       ;;
     -y | --yes)
       YES=1
+      FLAG=-y
       shift
       ;;
     -e | --env)
@@ -49,18 +53,262 @@ while :; do
   esac
 done
 
+if [ -n "${script_source}" ]; then
+  script_name="$(basename "${script_source}")"
+else
+  script_name="netdata-uninstaller.sh"
+fi
+
+info() {
+  echo >&2 "$(date) : INFO: ${script_name}: " "${1}"
+}
+
+error() {
+  echo >&2 "$(date) : ERROR: ${script_name}: " "${1}"
+  if [ -n "${NETDATA_SAVE_WARNINGS}" ]; then
+    NETDATA_WARNINGS="${NETDATA_WARNINGS}\n  - ${1}"
+  fi
+}
+
+fatal() {
+  echo >&2 "$(date) : FATAL: ${script_name}: FAILED TO UNINSTALL NETDATA: " "${1}"
+  if [ -n "${NETDATA_SAVE_WARNINGS}" ]; then
+    NETDATA_WARNINGS="${NETDATA_WARNINGS}\n  - ${1}"
+  fi
+  exit_reason "${1}" "${2}"
+  exit 1
+}
+
+exit_reason() {
+  if [ -n "${NETDATA_SAVE_WARNINGS}" ]; then
+    EXIT_REASON="${1}"
+    EXIT_CODE="${2}"
+    if [ -n "${NETDATA_PROPAGATE_WARNINGS}" ]; then
+      export EXIT_REASON
+      export EXIT_CODE
+      export NETDATA_WARNINGS
+    fi
+  fi
+}
+
 if [ "$YES" != "1" ]; then
   echo >&2 "This script will REMOVE netdata from your system."
   echo >&2 "Run it again with --yes to do it."
+  exit_reason "User did not accept uninstalling." R0001
   exit 1
 fi
 
-if [[ $EUID -ne 0 ]]; then
-  echo >&2 "This script SHOULD be run as root or otherwise it won't delete all installed components."
+if [ "$(id -u)" -ne 0 ]; then
+  error "This script SHOULD be run as root or otherwise it won't delete all installed components."
   key="n"
-  read -r -s -n 1 -p "Do you want to continue as non-root user [y/n] ? " key
+  read -r 1 -p "Do you want to continue as non-root user [y/n] ? " key
   if [ "$key" != "y" ] && [ "$key" != "Y" ]; then
+    exit_reason "User cancelled uninstall." R0002
     exit 1
+  fi
+fi
+
+user_input() {
+  if [ "${INTERACTIVITY}" = "-i" ]; then
+    TEXT="$1 [y/n]"
+
+    while true; do
+      echo "$TEXT"
+      read -r yn
+
+      case "$yn" in
+         [Yy]*) return 0;;
+         [Nn]*) return 1;;
+         *) echo "Please answer yes or no.";;
+       esac
+     done
+  fi
+}
+
+_cannot_use_tmpdir() {
+  testfile="$(TMPDIR="${1}" mktemp -q -t netdata-test.XXXXXXXXXX)"
+  ret=0
+
+  if [ -z "${testfile}" ]; then
+    return "${ret}"
+  fi
+
+  if printf '#!/bin/sh\necho SUCCESS\n' > "${testfile}"; then
+    if chmod +x "${testfile}"; then
+      if [ "$("${testfile}")" = "SUCCESS" ]; then
+        ret=1
+      fi
+    fi
+  fi
+
+  rm -f "${testfile}"
+  return "${ret}"
+}
+
+create_tmp_directory() {
+  if [ -z "${TMPDIR}" ] || _cannot_use_tmpdir "${TMPDIR}"; then
+    if _cannot_use_tmpdir /tmp; then
+      if _cannot_use_tmpdir "${PWD}"; then
+        fatal "Unable to find a usable temporary directory. Please set \$TMPDIR to a path that is both writable and allows execution of files and try again." R0003
+      else
+        TMPDIR="${PWD}"
+      fi
+    else
+      TMPDIR="/tmp"
+    fi
+  fi
+
+  mktemp -d -t netdata-kickstart-XXXXXXXXXX
+}
+
+tmpdir="$(create_tmp_directory)"
+
+detect_existing_install() {
+  if pkg_installed netdata; then
+    ndprefix="/"
+  else
+    if [ -n "${INSTALL_PREFIX}" ]; then
+      searchpath="${INSTALL_PREFIX}/bin:${INSTALL_PREFIX}/sbin:${INSTALL_PREFIX}/usr/bin:${INSTALL_PREFIX}/usr/sbin:${PATH}"
+      searchpath="${INSTALL_PREFIX}/netdata/bin:${INSTALL_PREFIX}/netdata/sbin:${INSTALL_PREFIX}/netdata/usr/bin:${INSTALL_PREFIX}/netdata/usr/sbin:${searchpath}"
+    else
+      searchpath="${PATH}"
+    fi
+
+    ndpath="$(PATH="${searchpath}" command -v netdata 2>/dev/null)"
+
+    if [ -z "$ndpath" ] && [ -x /opt/netdata/bin/netdata ]; then
+      ndpath="/opt/netdata/bin/netdata"
+    fi
+
+    if [ -n "${ndpath}" ]; then
+      ndprefix="$(dirname "$(dirname "${ndpath}")")"
+    fi
+
+    if echo "${ndprefix}" | grep -Eq '/usr$'; then
+      ndprefix="$(dirname "${ndprefix}")"
+    fi
+  fi
+
+  if [ -n "${ndprefix}" ]; then
+    typefile="${ndprefix}/etc/netdata/.install-type"
+    envfile="${ndprefix}/etc/netdata/.environment"
+    if [ -r "${typefile}" ]; then
+      ${ROOTCMD} sh -c "cat \"${typefile}\" > \"${tmpdir}/install-type\""
+      # shellcheck disable=SC1090,SC1091
+      . "${tmpdir}/install-type"
+    else
+      INSTALL_TYPE="unknown"
+    fi
+
+    if [ "${INSTALL_TYPE}" = "unknown" ] || [ "${INSTALL_TYPE}" = "custom" ]; then
+      if [ -r "${envfile}" ]; then
+        ${ROOTCMD} sh -c "cat \"${envfile}\" > \"${tmpdir}/environment\""
+        # shellcheck disable=SC1091
+        . "${tmpdir}/environment"
+        if [ -n "${NETDATA_IS_STATIC_INSTALL}" ]; then
+          if [ "${NETDATA_IS_STATIC_INSTALL}" = "yes" ]; then
+            INSTALL_TYPE="legacy-static"
+          else
+            INSTALL_TYPE="legacy-build"
+          fi
+        fi
+      fi
+    fi
+  fi
+}
+
+pkg_installed() {
+  case "${DISTRO_COMPAT_NAME}" in
+    debian|ubuntu)
+      dpkg-query --show --showformat '${Status}' "${1}" 2>&1 | cut -f 1 -d ' ' | grep -q '^install$'
+      return $?
+      ;;
+    centos|fedora|opensuse|ol)
+      rpm -q "${1}" > /dev/null 2>&1
+      return $?
+      ;;
+    *)
+      return 1
+      ;;
+  esac
+}
+
+detect_existing_install
+
+if [ -x "$(command -v apt-get)" ] && [ "${INSTALL_TYPE}" = "binpkg-deb" ]; then
+  if dpkg -s netdata > /dev/null; then
+    echo "Found netdata native installation"
+    if user_input "Do you want to remove netdata? "; then
+      apt-get remove netdata ${FLAG}
+    fi
+    if dpkg -s netdata-repo-edge > /dev/null; then
+      if user_input "Do you want to remove netdata-repo-edge? "; then
+        apt-get remove netdata-repo-edge ${FLAG}
+      fi
+    fi
+    if dpkg -s netdata-repo > /dev/null; then
+      if user_input "Do you want to remove netdata-repo? "; then
+        apt-get remove netdata-repo ${FLAG}
+      fi
+    fi
+    exit 0
+  fi
+elif [ -x "$(command -v dnf)" ] && [ "${INSTALL_TYPE}" = "binpkg-rpm" ]; then
+  if rpm -q netdata > /dev/null; then
+    echo "Found netdata native installation."
+    if user_input "Do you want to remove netdata? "; then
+      dnf remove netdata ${FLAG}
+    fi
+    if rpm -q netdata-repo-edge > /dev/null; then
+      if user_input "Do you want to remove netdata-repo-edge? "; then
+        dnf remove netdata-repo-edge ${FLAG}
+      fi
+    fi
+    if rpm -q netdata-repo > /dev/null; then
+      if user_input "Do you want to remove netdata-repo? "; then
+        dnf remove netdata-repo ${FLAG}
+      fi
+    fi
+    exit 0
+  fi
+elif [ -x "$(command -v yum)" ] && [ "${INSTALL_TYPE}" = "binpkg-rpm" ]; then
+  if rpm -q netdata > /dev/null; then
+    echo "Found netdata native installation."
+    if user_input "Do you want to remove netdata? "; then
+      yum remove netdata ${FLAG}
+    fi
+    if rpm -q netdata-repo-edge > /dev/null; then
+      if user_input "Do you want to remove netdata-repo-edge? "; then
+        yum remove netdata-repo-edge ${FLAG}
+      fi
+    fi
+    if rpm -q netdata-repo > /dev/null; then
+      if user_input "Do you want to remove netdata-repo? "; then
+        yum remove netdata-repo ${FLAG}
+      fi
+    fi
+    exit 0
+  fi
+elif [ -x "$(command -v zypper)" ] && [ "${INSTALL_TYPE}" = "binpkg-rpm" ]; then
+  if [ "${FLAG}" = "-y" ]; then
+    FLAG=-n
+  fi
+  if zypper search -i netdata > /dev/null; then
+    echo "Found netdata native installation."
+    if user_input "Do you want to remove netdata? "; then
+      zypper ${FLAG} remove netdata
+    fi
+    if zypper search -i netdata-repo-edge > /dev/null; then
+      if user_input "Do you want to remove netdata-repo-edge? "; then
+        zypper ${FLAG} remove netdata-repo-edge
+      fi
+    fi
+    if zypper search -i netdata-repo > /dev/null; then
+      if user_input "Do you want to remove netdata-repo? "; then
+        zypper ${FLAG} remove netdata-repo
+      fi
+    fi
+    exit 0
   fi
 fi
 
@@ -72,7 +320,8 @@ rcservice_cmd="$(command -v rc-service 2> /dev/null)"
 systemctl_cmd="$(command -v systemctl 2> /dev/null)"
 service() {
 
-  local cmd="${1}" action="${2}"
+  cmd="${1}"
+  action="${2}"
 
   if [ -n "${systemctl_cmd}" ]; then
     run "${systemctl_cmd}" "${action}" "${cmd}"
@@ -118,21 +367,13 @@ setup_terminal() {
 }
 setup_terminal || echo > /dev/null
 
-run_ok() {
-  printf >&2 "%s OK %s\n\n" "${TPUT_BGGREEN}${TPUT_WHITE}${TPUT_BOLD}" "${TPUT_RESET}"
-}
-
-run_failed() {
-  printf >&2 "%s FAILED %s\n\n" "${TPUT_BGRED}${TPUT_WHITE}${TPUT_BOLD}" "${TPUT_RESET}"
-}
-
 ESCAPED_PRINT_METHOD=
-if printf "%q " test > /dev/null 2>&1; then
+if printf "%s " test > /dev/null 2>&1; then
   ESCAPED_PRINT_METHOD="printfq"
 fi
 escaped_print() {
   if [ "${ESCAPED_PRINT_METHOD}" = "printfq" ]; then
-    printf "%q " "${@}"
+    printf "%s " "${@}"
   else
     printf "%s" "${*}"
   fi
@@ -141,9 +382,10 @@ escaped_print() {
 
 run_logfile="/dev/null"
 run() {
-  local user="${USER--}" dir="${PWD}" info info_console
+  user="${USER--}"
+  dir="${PWD}"
 
-  if [ "${UID}" = "0" ]; then
+  if [ "$(id -u)" = "0" ]; then
     info="[root ${dir}]# "
     info_console="[${TPUT_DIM}${dir}${TPUT_RESET}]# "
   else
@@ -163,12 +405,13 @@ run() {
 
   "${@}"
 
-  local ret=$?
+  ret=$?
   if [ ${ret} -ne 0 ]; then
-    run_failed
+    printf >&2 "%s FAILED %s\n\n" "${TPUT_BGRED}${TPUT_WHITE}${TPUT_BOLD}" "${TPUT_RESET}"
     printf >> "${run_logfile}" "FAILED with exit code %s\n" "${ret}"
+    NETDATA_WARNINGS="${NETDATA_WARNINGS}\n  - Command \"${*}\" failed with exit code ${ret}."
   else
-    run_ok
+    printf >&2 "%s OK %s\n\n" "${TPUT_BGGREEN}${TPUT_WHITE}${TPUT_BOLD}" "${TPUT_RESET}"
     printf >> "${run_logfile}" "OK\n"
   fi
 
@@ -176,17 +419,17 @@ run() {
 }
 
 portable_del_group() {
-  local groupname="${1}"
+  groupname="${1}"
 
   # Check if group exist
-  echo >&2 "Removing ${groupname} user group ..."
+  info "Removing ${groupname} user group ..."
 
   # Linux
   if command -v groupdel 1> /dev/null 2>&1; then
     if grep -q "${groupname}" /etc/group; then
       run groupdel "${groupname}" && return 0
     else
-      echo >&2 "Group ${groupname} already removed in a previous step."
+      info "Group ${groupname} already removed in a previous step."
       return 0
     fi
   fi
@@ -196,17 +439,21 @@ portable_del_group() {
     if dseditgroup -o read netdata 1> /dev/null 2>&1; then
       run dseditgroup -o delete "${groupname}" && return 0
     else
-      echo >&2 "Could not find group ${groupname}, nothing to do"
+      info "Could not find group ${groupname}, nothing to do"
       return 0
     fi
   fi
 
-  echo >&2 "Group ${groupname} was not automatically removed, you might have to remove it manually"
+  error "Group ${groupname} was not automatically removed, you might have to remove it manually"
   return 1
 }
 
 issystemd() {
-  local pids p myns ns systemctl
+  pids=''
+  p=''
+  myns=''
+  ns=''
+  systemctl=''
 
   # if the directory /lib/systemd/system OR /usr/lib/systemd/system (SLES 12.x) does not exit, it is not systemd
   if [ ! -d /lib/systemd/system ] && [ ! -d /usr/lib/systemd/system ]; then
@@ -240,8 +487,8 @@ issystemd() {
 }
 
 portable_del_user() {
-  local username="${1}"
-  echo >&2 "Deleting ${username} user account ..."
+  username="${1}"
+  info "Deleting ${username} user account ..."
 
   # Linux
   if command -v userdel 1> /dev/null 2>&1; then
@@ -253,15 +500,16 @@ portable_del_user() {
     run sysadminctl -deleteUser "${username}" && return 0
   fi
 
-  echo >&2 "User ${username} could not be deleted from system, you might have to remove it manually"
+  error "User ${username} could not be deleted from system, you might have to remove it manually"
   return 1
 }
 
 portable_del_user_from_group() {
-  local groupname="${1}" username="${2}"
+  groupname="${1}"
+  username="${2}"
 
   # username is not in group
-  echo >&2 "Deleting ${username} user from ${groupname} group ..."
+  info "Deleting ${username} user from ${groupname} group ..."
 
   # Linux
   if command -v gpasswd 1> /dev/null 2>&1; then
@@ -283,43 +531,38 @@ portable_del_user_from_group() {
     run dseditgroup -o delete -u "${username}" "${groupname}" && return 0
   fi
 
-  echo >&2 "Failed to delete user ${username} from group ${groupname} !"
+  error "Failed to delete user ${username} from group ${groupname} !"
   return 1
 }
 
 quit_msg() {
   echo
   if [ "$FILE_REMOVAL_STATUS" -eq 0 ]; then
-    echo >&2 "Something went wrong :("
+    fatal "Failed to completely remove Netdata from this system." R0004
   else
-    echo >&2 "Netdata files were successfully removed from your system"
-  fi
-}
-
-user_input() {
-  TEXT="$1"
-  if [ "${INTERACTIVITY}" = "-i" ]; then
-    read -r -p "$TEXT" >&2
+    info "Netdata files were successfully removed from your system"
   fi
 }
 
 rm_file() {
   FILE="$1"
   if [ -f "${FILE}" ]; then
-    run rm -v ${INTERACTIVITY} "${FILE}"
+    if user_input "Do you want to delete this file '$FILE' ? "; then
+      run rm -v "${FILE}"
+    fi
   fi
 }
 
 rm_dir() {
   DIR="$1"
   if [ -n "$DIR" ] && [ -d "$DIR" ]; then
-    user_input "Press ENTER to recursively delete directory '$DIR' > "
-    run rm -v -f -R "${DIR}"
+    if user_input "Do you want to delete this directory '$DIR' ? "; then
+      run rm -v -f -R "${DIR}"
+    fi
   fi
 }
 
 safe_pidof() {
-  local pidof_cmd
   pidof_cmd="$(command -v pidof 2> /dev/null)"
   if [ -n "${pidof_cmd}" ]; then
     ${pidof_cmd} "${@}"
@@ -345,14 +588,16 @@ pidisnetdata() {
 }
 
 stop_netdata_on_pid() {
-  local pid="${1}" ret=0 count=0
+  pid="${1}"
+  ret=0
+  count=0
 
   pidisnetdata "${pid}" || return 0
 
-  printf >&2 "Stopping netdata on pid %s ..." "${pid}"
+  info "Stopping netdata on pid ${pid} ..."
   while [ -n "$pid" ] && [ ${ret} -eq 0 ]; do
     if [ ${count} -gt 24 ]; then
-      echo >&2 "Cannot stop the running netdata on pid ${pid}."
+      error "Cannot stop the running netdata on pid ${pid}."
       return 1
     fi
 
@@ -377,16 +622,17 @@ stop_netdata_on_pid() {
 
   echo >&2
   if [ ${ret} -eq 0 ]; then
-    echo >&2 "SORRY! CANNOT STOP netdata ON PID ${pid} !"
+    error "SORRY! CANNOT STOP netdata ON PID ${pid} !"
     return 1
   fi
 
-  echo >&2 "netdata on pid ${pid} stopped."
+  info "netdata on pid ${pid} stopped."
   return 0
 }
 
 netdata_pids() {
-  local p myns ns
+  p=''
+  ns=''
 
   myns="$(readlink /proc/self/ns/pid 2> /dev/null)"
 
@@ -403,49 +649,57 @@ netdata_pids() {
 }
 
 stop_all_netdata() {
-  local p
+  p=''
+  stop_success=0
 
-  if [ "${UID}" -eq 0 ]; then
+  if [ "$(id -u)" -eq 0 ]; then
     uname="$(uname 2> /dev/null)"
 
     # Any of these may fail, but we need to not bail if they do.
     if issystemd; then
       if systemctl stop netdata; then
+        stop_success=1
         sleep 5
       fi
     elif [ "${uname}" = "Darwin" ]; then
       if launchctl stop netdata; then
+        stop_success=1
         sleep 5
       fi
     elif [ "${uname}" = "FreeBSD" ]; then
       if /etc/rc.d/netdata stop; then
+        stop_success=1
         sleep 5
       fi
     else
       if service netdata stop; then
+        stop_success=1
         sleep 5
       fi
     fi
   fi
 
-  if [ -n "$(netdata_pids)" ] && [ -n "$(builtin type -P netdatacli)" ]; then
-    netdatacli shutdown-agent
-    sleep 20
-  fi
+  if [ "$stop_success" = "0" ]; then
+    if [ -n "$(netdata_pids)" ] && [ -n "$(command -v netdatacli)" ]; then
+      netdatacli shutdown-agent
+      sleep 20
+    fi
 
-  for p in $(netdata_pids); do
-    # shellcheck disable=SC2086
-    stop_netdata_on_pid ${p}
-  done
+    for p in $(netdata_pids); do
+      # shellcheck disable=SC2086
+      stop_netdata_on_pid ${p}
+    done
+  fi
 }
 
 trap quit_msg EXIT
 
-#shellcheck source=/dev/null
-source "${ENVIRONMENT_FILE}" || exit 1
+# shellcheck source=/dev/null
+# shellcheck disable=SC1090
+. "${ENVIRONMENT_FILE}" || exit 1
 
 #### STOP NETDATA
-echo >&2 "Stopping a possibly running netdata..."
+info "Stopping a possibly running netdata..."
 stop_all_netdata
 
 #### REMOVE NETDATA FILES
@@ -484,16 +738,19 @@ FILE_REMOVAL_STATUS=1
 
 #### REMOVE NETDATA USER FROM ADDED GROUPS
 if [ -n "$NETDATA_ADDED_TO_GROUPS" ]; then
-  user_input "Press ENTER to delete 'netdata' from following groups: '$NETDATA_ADDED_TO_GROUPS' > "
-  for group in $NETDATA_ADDED_TO_GROUPS; do
-    portable_del_user_from_group "${group}" "netdata"
-  done
+  if user_input "Do you want to delete 'netdata' from following groups: '$NETDATA_ADDED_TO_GROUPS' ? "; then
+    for group in $NETDATA_ADDED_TO_GROUPS; do
+      portable_del_user_from_group "${group}" "netdata"
+    done
+  fi
 fi
 
 #### REMOVE USER
-user_input "Press ENTER to delete 'netdata' system user > "
-portable_del_user "netdata" || :
+if user_input "Do you want to delete 'netdata' system user ? "; then
+  portable_del_user "netdata" || :
+fi
 
 ### REMOVE GROUP
-user_input "Press ENTER to delete 'netdata' system group > "
-portable_del_group "netdata" || :
+if user_input "Do you want to delete 'netdata' system group ? "; then
+  portable_del_group "netdata" || :
+fi
